NotificationService: Add notification for new test results
Interesting: EF Core, upon `Add`, gives entities a temporary ID, which is replaced upon `SaveChanges`. This means that you don't have to do a roundtrip to the database to get an ID generated by the database server. Use the temporary ID provided by EF Core and it'll work out to the same result once the transaction is complete.